moved all Double values to Float values because of http://www.exploringbinary.com/java-hangs-when-converting-2-2250738585072012e-308/ YaCy does not really need double-precision floating point computation anywhere, so this should not affect any feature
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@7460 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ htroot/ConfigUpdate_p.java
@@ -156,10 +156,10 @@ public class ConfigUpdate_p {
         final String versionstring = yacyBuildProperties.getVersion() + "/" + yacyBuildProperties.getSVNRevision();
         prop.putHTML("candeploy_versionpp", versionstring);
         final boolean devenvironment = new File(sb.getAppPath(), ".svn").exists();
-        double thisVersion = Double.parseDouble(yacyBuildProperties.getVersion());
+        float thisVersion = Float.parseFloat(yacyBuildProperties.getVersion());
         // cut off the SVN Rev in the Version
         try {
-            thisVersion = Math.round(thisVersion*1000.0)/1000.0;
+            thisVersion = (float) (Math.round(thisVersion*1000.0)/1000.0);
         } catch (final NumberFormatException e) {}
