+++ htroot/osm.java
@@ -15,15 +15,15 @@ public class osm {
     public static RasterPlotter respond(final RequestHeader header, final serverObjects post, final serverSwitch env) {
 
         int zoom = 10;
-        double lat = 50.11670d;
-        double lon = 8.68333d;
+        float lat = 50.11670f;
+        float lon = 8.68333f;
         int width = 3;
         int height = 3;
         
         if (post != null) {
             zoom = post.getInt("zoom", zoom);
-            lat = post.getDouble("lat", lat);
-            lon = post.getDouble("lon", lon);
+            lat = post.getFloat("lat", lat);
+            lon = post.getFloat("lon", lon);
             width = post.getInt("width", width);
             height = post.getInt("height", height);
         }
